requirements.txt
mutagen==1.47.0
requests==2.32.3
python-dotenv==1.0.1

